completed POS application following coding standards

+++ src/main/java/com/ust/pos/user/UserController.java
@@ -4,6 +4,8 @@ import com.ust.pos.dto.UserDto;
 import com.ust.pos.role.service.RoleService;
 import com.ust.pos.user.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
@@ -45,8 +47,16 @@ public class UserController {
     }
 
     @GetMapping("/delete")
-    public String delete(@RequestParam String username) {
-        userService.delete(username);
+    public String delete(Model model, @RequestParam String username) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if(authentication != null){
+            String loggedInUser = authentication.getName();
+            userService.delete(username);
+            if (loggedInUser.equals(username)) {
+                SecurityContextHolder.clearContext();
+                return "redirect:/login";
+            }
+        }
         return "redirect:/user/list";
     }
 }
